+++ tree_unique_args/src/main.rs
@@ -14,6 +14,7 @@ pub(crate) mod subst;
 pub(crate) mod switch_rewrites;
 pub(crate) mod simple;
 pub(crate) mod util;
+pub(crate) mod type_analysis;
 
 pub type Result = std::result::Result<(), egglog::Error>;
